the booter update
The Corolla E100 is the seventh generation of cars sold by Toyota under the Corolla nameplate. This generation of Corolla is larger, heavier, and visually more aerodynamic than the model it replaced. With its 2,465 mm (97.0 in) wheelbase, the Corolla had moved into the compact size class once occupied by the Corona and Camry. The Corolla again had an equivalent model Sprinter, with the Sprinter Trueno being equivalent to the Corolla Levin and both exclusive to Toyota Vista Store Japanese dealerships.

The development of the seventh-generation Corolla was led by chief engineer Akihiko Saito.[7] Not only was the wheelbase increased, but the new Corolla also received a wider track than did the 90-series.[8] The solid design reflected the desire of development chief Dr. Akihiko Saito to make a 'mini-Lexus', to build on the success of Toyota's flagship range. The sedan and hatchback have drag coefficients of Cd=0.33, the liftback is Cd=0.30, and the wagon/van have Cd=0.36.[9] The model used fewer body panels than the predecessor, for increased strength, lower cost, and fewer panel gaps (providing a cleaner appearance and lower wind resistance). The headlights were now from polycarbonate rather than glass, and almost 90 percent of the bodyshell was made from galvanized steel - compared to about 60 percent in the preceding generation.[10] 

https://en.wikipedia.org/wiki/Toyota_Corolla_(E100)